third_party/blink/web_tests/platform/generic/external/wpt/fetch/http-cache/invalidate.any.serviceworker-expected.txt

This is a testharness.js-based test.
PASS HTTP cache invalidates after a successful response from a POST
PASS HTTP cache does not invalidate after a failed response from an unsafe request
PASS HTTP cache invalidates after a successful response from a PUT
PASS HTTP cache invalidates after a successful response from a DELETE
FAIL HTTP cache invalidates after a successful response from an unknown method assert_equals: Response 3 comes from cache expected 3 but got 1
FAIL HTTP cache invalidates Location URL after a successful response from a POST assert_equals: Response 3 comes from cache expected 3 but got 1
PASS HTTP cache does not invalidate Location URL after a failed response from an unsafe request
FAIL HTTP cache invalidates Location URL after a successful response from a PUT assert_equals: Response 3 comes from cache expected 3 but got 1
FAIL HTTP cache invalidates Location URL after a successful response from a DELETE assert_equals: Response 3 comes from cache expected 3 but got 1
FAIL HTTP cache invalidates Location URL after a successful response from an unknown method assert_equals: Response 3 comes from cache expected 3 but got 1
FAIL HTTP cache invalidates Content-Location URL after a successful response from a POST assert_equals: Response 3 comes from cache expected 3 but got 1
PASS HTTP cache does not invalidate Content-Location URL after a failed response from an unsafe request
FAIL HTTP cache invalidates Content-Location URL after a successful response from a PUT assert_equals: Response 3 comes from cache expected 3 but got 1
FAIL HTTP cache invalidates Content-Location URL after a successful response from a DELETE assert_equals: Response 3 comes from cache expected 3 but got 1
FAIL HTTP cache invalidates Content-Location URL after a successful response from an unknown method assert_equals: Response 3 comes from cache expected 3 but got 1
Harness: the test ran to completion.
